fix: get packet timestamps before logging call
The previous version had a bug because the timestamp was being taken in
the function call where *all* packets were getting the same timestamp.
Apparently the `aio.TaskGroup()` was doing some kind of optimization
that we ran up against.

Also ensures that all `datetime.now()` calls get the local timezone
attached properly so all timestamps are written in system-local time
instead of some in UTC.