Timestamps and Broker Restart fixups (#4)
* fix: get packet timestamps before logging call

The previous version had a bug because the timestamp was being taken in
the function call where *all* packets were getting the same timestamp.
Apparently the `aio.TaskGroup()` was doing some kind of optimization
that we ran up against.

Also ensures that all `datetime.now()` calls get the local timezone
attached properly so all timestamps are written in system-local time
instead of some in UTC.

* Archiving safety improvements

Switches to vanilla Info-Zip and adds some resources limits to the
systemd services to try to prevent crashes.

* fix: Dump in-progress json files before Client exit during unwind

Apparently, probably due to python's magic CoW-reference semantics,
dumping the in-progress sensor data requires the MQTT client object to
still be around. So, put the data flush call above the Client teardown
in the stack unwind. Should fix #2.

* fix: Use system-wide `mosquitto` instance instead of custom `mqtt-pignet`

Helps to resolve #2, since decoupling the MQTT broker from the `pignet`
data logger should reduce how often the broker is restarting, which in
turn is leading to a lot of issues in the data logger. In general, I
think we should avoid restarting `mosquitto` unless it's definitely
necessary.